feat: create errors implementation

+++ server/src/modules/students/use-cases/delete-student/delete-student-use-case.ts
@@ -1,3 +1,4 @@
+import { NotFoundError } from '../../../../errors/not-found-error'
 import { IStudentsRepository } from '../../repositories/students-repository.interface'
 import { IUseCase } from '../use-case.interface'
 
@@ -5,6 +6,9 @@ class DeleteStudentUseCase implements IUseCase<string, void> {
     constructor(private readonly studentsRepository: IStudentsRepository) {}
 
     async execute(id: string): Promise<void> {
+        const student = await this.studentsRepository.getById(id)
+        if (!student) throw new NotFoundError('Student not found')
+
         await this.studentsRepository.delete(id)
     }
 }
